project folders restructure
