+++ app.js
@@ -1,20 +1,11 @@
 require('dotenv').config()
+const bot = require('./config/botConfig')
+const { getWeatherData } = require('./helpers/weather')
+const { getTimeData } = require('./helpers/time')
+const { getUserData, resetUserData } = require('./helpers/userData')
 
-const TelegramBot = require('node-telegram-bot-api')
-const axios = require('axios')
-const moment = require('moment-timezone')
-
-const TELEGRAM_BOT_TOKEN = process.env.TELEGRAM_BOT_TOKEN
-const OPENWEATHERMAP_API_KEY = process.env.OPENWEATHERMAP_API_KEY
-
-// replace the value below with the Telegram token you receive from @BotFather
-const token = TELEGRAM_BOT_TOKEN
-
-// Create a bot instance
-const bot = new TelegramBot(token, { polling: true })
-
-// Listen for the /start command
-bot.onText(/\/start/, (msg) => {
+// Listen for the start command
+bot.onText(/^start$/, (msg) => {
   const chatId = msg.chat.id
   bot.sendMessage(
     chatId,
@@ -75,79 +66,15 @@ bot.on('message', async (msg) => {
 })
 
 // Helper function to get weather data for a city
-async function getWeatherData(city) {
-  // Log the request to retrieve weather data
-  console.log(`Requesting weather data for ${city}...`)
-
-  // Get the weather data for the requested city from the OpenWeatherMap API
-  const response = await axios.get(
-    `http://api.openweathermap.org/data/2.5/weather?q=${city}&appid=${OPENWEATHERMAP_API_KEY}`
-  )
-  const weatherData = response.data
-
-  // Get the current time for the requested city
-  const timezone = moment.tz.guess()
-  const time = moment().tz(timezone).format('h:mm A')
-
-  // Log the retrieved weather data
-  console.log(
-    `Weather data retrieved: ${JSON.stringify(weatherData)}, Time: ${time}`
-  )
-
-  // Format the weather data as a string message
-  const weatherDescription = weatherData.weather[0].description
-  const temperature = Math.round(weatherData.main.temp - 273.15)
-  const messageText = `The weather in ${city} is currently ${weatherDescription} with a temperature of ${temperature}°C. The current time is ${time}.`
-  // Log the weather data message sent to the user
-  console.log(`Weather data message sent to user: ${messageText}`)
-
-  // Return the weather data message
-  return messageText
-}
 
 // Helper function to get time data for a city
-async function getTimeData(city) {
-  // Get the current time for the requested city
-  const timezone = moment.tz.guess()
-  const time = moment().tz(timezone).format('h:mm A')
-
-  // Log the retrieved time data
-  console.log(`Time data retrieved: Time: ${time}`)
-
-  // Format the time data as a string message
-  const messageText = `The current time in ${city} is ${time}.`
-
-  // Log the time data message sent to the user
-  console.log(`Time data message sent to user: ${messageText}`)
-
-  // Return the time data message
-  return messageText
-}
 
 // Helper function to get user data from storage or create new data if it doesn't exist
-function getUserData(chatId) {
-  let userData = storage[chatId]
-  if (!userData) {
-    userData = {
-      waitingForCity: false,
-      waitingForWeather: false,
-      waitingForTime: false,
-    }
-    storage[chatId] = userData
-  }
-  return userData
-}
 
 // Helper function to reset user data
-function resetUserData(chatId) {
-  const userData = getUserData(chatId)
-  userData.waitingForCity = false
-  userData.waitingForWeather = false
-  userData.waitingForTime = false
-}
 
 // Listen for the /stop command
-bot.onText(/\/stop/, (msg) => {
+bot.onText(/stop/, (msg) => {
   const chatId = msg.chat.id
   bot.sendMessage(chatId, 'You have stopped the current operation.')
 
@@ -158,22 +85,22 @@ bot.onText(/\/stop/, (msg) => {
 // Initialize storage
 const storage = {}
 
-// Listen for the /start command
-bot.onText(/\/start/, (msg) => {
-  const chatId = msg.chat.id
-  bot.sendMessage(
-    chatId,
-    'Hello! This bot can show you the weather and time for any city. To use it, please choose an option below:',
-    {
-      reply_markup: {
-        inline_keyboard: [
-          [{ text: 'Get Weather', callback_data: 'get_weather' }],
-          [{ text: 'Get Time', callback_data: 'get_time' }],
-        ],
-      },
-    }
-  )
-})
+// Listen for the start command
+// bot.onText(/start/, (msg) => {
+//   const chatId = msg.chat.id
+//   bot.sendMessage(
+//     chatId,
+//     'Hello! This bot can show you the weather and time for any city. To use it, please choose an option below:',
+//     {
+//       reply_markup: {
+//         inline_keyboard: [
+//           [{ text: 'Get Weather', callback_data: 'get_weather' }],
+//           [{ text: 'Get Time', callback_data: 'get_time' }],
+//         ],
+//       },
+//     }
+//   )
+// })
 
 // Listen for inline keyboard button presses
 bot.on('callback_query', async (callbackQuery) => {
@@ -202,7 +129,7 @@ bot.on('callback_query', async (callbackQuery) => {
       // Ask user for the city name
       bot.sendMessage(
         chatId,
-        'Please enter the name of the city or send /stop to cancel:'
+        'Please enter the name of the city or send stop to cancel:'
       )
       break
     default:
